Tidied up the repository
I've merged some of the files together to make navigating
the repo a bit easier
