--- webenrollmentrequest.go
@@ -1,8 +1,0 @@
-package adcs
-
-// WebEnrollmentRequest interface is implemented by the
-// various kinds of
-type WebEnrollmentRequest interface {
-	Submit() (WebEnrollmentResponse, error)
-	GetServer() *WebEnrollmentServer
-}
